+++ src/services/proprietarios.service.js
@@ -2,13 +2,18 @@ import ProprietarioRepository from '../repositories/proprietario.repository.js'
 
 class ProprietarioService {
   async create({ nome, cpf, telefone }) {
-    const proprietarioJaCadastrado = await ProprietarioRepository.listByProprietario({ cpf })
+    try {
+      const proprietarioJaCadastrado = await ProprietarioRepository.listByProprietario({ cpf })
 
     if(proprietarioJaCadastrado) {
       throw new Error('Proprietário já existe, cpf já cadastrado!');
     }
 
     return await ProprietarioRepository.create({ nome, cpf, telefone })
+    } catch (error) {
+      throw error
+    }
+    
   }
 
   async list() {
@@ -21,9 +26,13 @@ class ProprietarioService {
 
   async listById({ proprietarioId }) {
     try {
-      const result = await ProprietarioRepository.listById({ proprietarioId });
+      const ownerExist = await ProprietarioRepository.listById({ proprietarioId });
 
-      return { proprietario: result}
+      if(!ownerExist) {
+        throw new Error('Proprietario não encontrado')
+      } 
+
+      return { Proprietario:  ownerExist}
     } catch (error) {
       throw error
     }
@@ -31,7 +40,12 @@ class ProprietarioService {
 
   async listPetsProprietario({ proprietarioId }) {
     try {
-      await ProprietarioRepository.listById({ proprietarioId });
+      const ownerExist = await ProprietarioRepository.listById({ proprietarioId });
+
+      if(!ownerExist) {
+        throw new Error('Proprietario não encontrado')
+      } 
+
       return await ProprietarioRepository.listPetsProprietario({ proprietarioId })
 
     } catch (error) {
@@ -41,7 +55,12 @@ class ProprietarioService {
 
   async update({ nome, telefone, proprietarioId }) {
     try {
-      await ProprietarioRepository.listById({ proprietarioId });
+      const ownerExist = await ProprietarioRepository.listById({ proprietarioId });
+
+      if(!ownerExist) {
+        throw new Error('Proprietario não encontrado')
+      } 
+
       return await ProprietarioRepository.update({ nome, telefone, proprietarioId })
     } catch (error) {
       throw error
@@ -50,7 +69,18 @@ class ProprietarioService {
 
   async delete({ proprietarioId }) {
     try {
-      await ProprietarioRepository.listById({ proprietarioId })
+      const ownerExist = await ProprietarioRepository.listById({ proprietarioId })
+      
+      if(!ownerExist) {
+        throw new Error('Proprietario não encontrado')
+      } 
+
+      const areTherePets = await ProprietarioRepository.listPetsProprietario({ proprietarioId })
+      
+      if (areTherePets.length > 0) {
+        throw new Error("Proprietário possui pets, delete-os primeiro")
+      }
+
       return await ProprietarioRepository.delete({ proprietarioId })
     } catch (error) {
       throw error
